// frontend/src/components/finance/pl/pl-cost-add.d.ts
import type { NewRouteFunction } from "../../../types/util-types/new-route.type";
import { CommonParams } from "../../../utils/common_params";
export declare class PlCostAdd {
    private readonly openNewRoute;
    commonParams: CommonParams | undefined;
    private readonly categoryElement;
    private readonly dateElement;
    private readonly amountElement;
    private readonly commentElement;
    private readonly commonErrorElement;
    private readonly validations;
    constructor(openNewRoute: NewRouteFunction, commonParams: CommonParams | undefined);
    private addTransact;
    private saveTransact;
    private initial;
    private showCategories;
    loadData(): Promise<void>;
}
//# sourceMappingURL=pl-cost-add.d.ts.map